make generic list page, move books/movies to use generic list
