+++ MediaSet.Remix/app/root.tsx
@@ -7,7 +7,7 @@ import {
   Scripts,
   ScrollRestoration,
 } from "@remix-run/react";
-import { LibraryBig } from "lucide-react";
+import { Clapperboard, LibraryBig } from "lucide-react";
 import PendingNavigation from "./components/pending-navigation";
 
 const year = new Date().getUTCFullYear();
@@ -28,6 +28,7 @@ export default function App() {
             <div className="flex flex-row gap-4 items-center">
               <NavLink to="/" className="p-3 flex items-center rounded-lg">Home</NavLink>
               <NavLink to="/books" className="p-3 flex gap-2 items-center rounded-lg"><LibraryBig /> Books</NavLink>
+              <NavLink to="/movies" className="p-3 flex gap-2 items-center rounded-lg"><Clapperboard /> Movies</NavLink>
             </div>
           </div>
           <PendingNavigation />
