+++ MediaSet.Remix/app/routes/books2.$bookId_.delete/route.tsx
@@ -1,12 +1,12 @@
 import type { ActionFunctionArgs } from "@remix-run/node";
 import { redirect } from "@remix-run/node";
 import invariant from "tiny-invariant"
-import { Entities } from "~/constants";
+import { Entity } from "~/constants";
 import { deleteEntity } from "~/entity-data";
 import { Book } from "~/models";
 
 export const action = async ({ params }: ActionFunctionArgs) => {
   invariant(params.bookId, "Missing bookId param");
-  await deleteEntity<Book>(Entities.Books, params.bookId);
+  await deleteEntity<Book>(Entity.Books, params.bookId);
   return redirect('/books');
 };
